Add requirements, fix Windows sound
Add requirements.txt files for both running and building the software.

As a byproduct of adding requirements.txt, fix sound on Windows by
pinning the scoring software to an older version of PyQt. The 6.7.x
series of releases have a regression that cause sound on Windows to not
work without manually adding ffmpeg DLLs to the package installation.