Prevent foghorn from sounding after match end buzzer (#32)
Commit 7d45cbf29f80f8b6064a703dedfea6077923cd16 introduced a regression
that causes the foghorn (match abort sound) to trigger after the match
complete buzzer. The root cause is that the timer is still considered
"running" when resetTimer is called (both by timerComplete and
redundantly in updateTimer).

Until the timer logic is moved into the main application (see #2), patch
this by explicitly setting the timer running variable to False before
doing anything that would reset the timer, and also remove the redundant
resetTimer call.