Introduce resource packs
Introduce the concept of an installable "resource pack" that contains
licensed assets. Migrate all usages of resources to point to the
installed resource pack, rather than the `res` directory.

This allows us to distribute a single executable for the application and
then install resources later, rather than having to maintain a separate
work tree to build licensed versions of the executables.