Add scoresheet entry
Add ability to enter scoresheets through the scoring software directly,
rather than calculating the score manually in another window and
entering the calculated score manually. An XML file defines the
clickable areas and values for each option within each task within each
mission. Once a complete scoresheet is entered, make a request to the
backend included in the resource pack to calculate the actual score,
then insert that score into the Team object, and save off the scoresheet
for later analysis.

This implementation fixes many of the issues with beta/work-in-progress
revisions of scoresheet entry:
* Scoresheet is a dialog, not a window, so the main scorekeeping window
  can be interacted with while a scoresheet is pulled up
* Scoresheet window is resizable
* Confirmation is shown before overwriting an existing score
* Match number is automatically selected based on what matches have been
  entered so far