Add keyboard entry method to scoresheet
Add the ability to use the keyboard to enter scores into the scoresheet
entry window. Enter is used to activate keyboard mode, and pressing
number keys will choose the corresponding option in the scoresheet
(including 0/1 for false/true choices). A line is drawn under the
currently-selected task, and the task selection auto-advances once a
selection is made.